+++ include/struct.h
@@ -28,12 +28,12 @@ typedef union s_tuple
 {
 	struct
 	{
-		float	x;
-		float	y;
-		float	z;
-		float	w;
+		double	x;
+		double	y;
+		double	z;
+		double	w;
 	};
-	float		tuple[4];
+	double		tuple[4];
 }	t_tuple;
 
 typedef struct s_ray
@@ -42,24 +42,24 @@ typedef struct s_ray
 	{
 		struct
 		{
-			float	x;
-			float	y;
-			float	z;
-			float	w;
+			double	x;
+			double	y;
+			double	z;
+			double	w;
 		};
-		float		tab[4];
+		double		tab[4];
 		t_tuple		vector;
 	};
 	union
 	{
 		struct
 		{
-			float	x2;
-			float	y2;
-			float	z2;
-			float	w2;
+			double	x2;
+			double	y2;
+			double	z2;
+			double	w2;
 		};
-		float		tab2[4];
+		double		tab2[4];
 		t_tuple		point;
 	};
 }	t_ray;
@@ -70,24 +70,24 @@ typedef struct s_light
 	{
 		struct
 		{
-			float	x;
-			float	y;
-			float	z;
-			float	w;
+			double	x;
+			double	y;
+			double	z;
+			double	w;
 		};
-		float		tab[4];
+		double		tab[4];
 		t_tuple		intensity;
 	};
 	union
 	{
 		struct
 		{
-			float	x2;
-			float	y2;
-			float	z2;
-			float	w2;
+			double	x2;
+			double	y2;
+			double	z2;
+			double	w2;
 		};
-		float		tab2[4];
+		double		tab2[4];
 		t_tuple		position;
 	};
 }	t_light;
@@ -95,11 +95,11 @@ typedef struct s_light
 typedef struct s_material
 {
 	t_tuple	color;
-	float	ambient;
+	double	ambient;
 	t_tuple	a_color;
-	float	diffuse;
-	float	specular;
-	float	shininess;
+	double	diffuse;
+	double	specular;
+	double	shininess;
 }	t_material;
 
 typedef union u_bgra
@@ -118,25 +118,25 @@ typedef union s_3f
 {
 	struct
 	{
-		float	a;
-		float	b;
-		float	c;
+		double	a;
+		double	b;
+		double	c;
 	};
-	float		f[3];
+	double		f[3];
 }	t_3f;
 
 typedef union s_6f
 {
 	struct
 	{
-		float	a;
-		float	b;
-		float	c;
-		float	d;
-		float	e;
-		float	f;
+		double	a;
+		double	b;
+		double	c;
+		double	d;
+		double	e;
+		double	f;
 	};
-	float		t[6];
+	double		t[6];
 }	t_6f;
 
 typedef struct s_ijkl
@@ -156,13 +156,13 @@ typedef struct s_ijkl
 /* ************************************************************************** */
 typedef struct s_camera
 {
-	float		hsize;
-	float		vsize;
-	float		field_of_view;
+	double		hsize;
+	double		vsize;
+	double		field_of_view;
 	t_matrix4	transform;
-	float		half_width;
-	float		half_height;
-	float		pixel_size;
+	double		half_width;
+	double		half_height;
+	double		pixel_size;
 }	t_camera;
 
 /* ************************************************************************** */
@@ -176,8 +176,8 @@ typedef struct s_object
 	t_tuple			position;
 	t_matrix4		transform;
 	t_material		material;
-	float			cyl_min;
-	float			cyl_max;
+	double			cyl_min;
+	double			cyl_max;
 	int				cyl_closed;
 	struct s_object	*prev;
 	struct s_object	*next;
@@ -207,7 +207,7 @@ typedef struct s_intersection
 
 typedef struct s_t_and_object
 {
-	float					t;
+	double					t;
 	t_object				*object;
 	struct s_t_and_object	*prev;
 	struct s_t_and_object	*next;
@@ -215,13 +215,13 @@ typedef struct s_t_and_object
 
 typedef struct s_data_key
 {
-	float	c_add_pos_x;
-	float	c_add_pos_y;
-	float	c_add_pos_z;
-	float	c_add_to_x;
-	float	c_add_to_y;
-	float	c_add_to_z;
-	float	c_add_fov;
+	double	c_add_pos_x;
+	double	c_add_pos_y;
+	double	c_add_pos_z;
+	double	c_add_to_x;
+	double	c_add_to_y;
+	double	c_add_to_z;
+	double	c_add_fov;
 	int		key_a;
 	int		key_w;
 	int		key_s;
@@ -233,13 +233,13 @@ typedef struct s_data_key
 
 typedef struct s_data_parsing
 {
-	float	ambient;
+	double	ambient;
 	t_tuple	a_color;
 	t_tuple	c_position;
 	t_tuple	c_to;
-	float	c_fov;
+	double	c_fov;
 	t_tuple	l_position;
-	float	l_i;
+	double	l_i;
 	t_tuple	l_color;
 }	t_dp;
