+++ source/world/color_at.c
@@ -6,7 +6,7 @@
 /*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 16:48:57 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/02 14:56:38 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/15 11:53:31 by tda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 // 	if (!inter)
 // 		return ((t_tuple){0});
 // 	t = hit_to(inter);
-// 	if (!t || t->t == 0 || t->t == FLT_MAX)
+// 	if (!t || t->t == 0 || t->t == DBL_MAX)
 // 	{
 // 		free(t);
 // 		return ((t_tuple){0});
@@ -42,7 +42,7 @@ t_tuple	color_at(t_world *w, t_ray r)
 	t_tuple	ret;
 
 	inter = intersect_world(w, r);
-	if (inter.t == FLT_MAX || !inter.object)
+	if (inter.t == DBL_MAX || !inter.object)
 		return ((t_tuple){0});
 	comps = prepare_computations(r, &inter);
 	ret = shade_hit(*w, comps);
